refactor(35447): export primary key wrapper
